+++ src/Table.java
@@ -1,22 +1,22 @@
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class Table extends HashMap<Integer, Ligne> {
+public class Table extends HashMap<Double, Ligne> {
 
     public Table() {
         super();
     }
 
-    public void ajouterLigne(int S, int... valeurs) {
+    public void ajouterLigne(double S, double... valeurs) {
         Ligne l = new Ligne(valeurs);
         this.put(S, l);
     }
 
-    public int getX(int s) {
+    public double getX(double s) {
         return this.get(s).getX();
     }
 
-    public int getF(int s){
+    public double getF(double s){
         return this.get(s).getF();
     }
